Turns out 'unsafeRunWhenOn' only protects the runtime. More accurately, all classes that are referenced directly in your startup code must be safe to load completely on all Dists. The simple solution is 'If a class contains client-only code, it goes in the client module.'
